requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
comm==0.2.2
coverage==7.6.9
debugpy==1.8.11
decorator==5.1.1
Django==4.2
djangorestframework==3.15.2
execnet==2.1.1
executing==2.1.0
Flask==3.1.0
freezegun==1.5.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
MechanicalSoup==1.3.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pipdeptree==2.24.0
platformdirs==4.3.6
playwright==1.49.1
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyee==12.0.0
Pygments==2.19.1
pyspark==3.5.4
pytest==8.3.4
pytest-cov==6.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
pyzmq==26.2.0
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
